third_party/blink/web_tests/external/wpt/css/css-pseudo/idlharness-expected.txt

This is a testharness.js-based test.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "TypeError: window.getPseudoElements is not a function"
[FAIL] CSSPseudoElement interface: existence and properties of interface object
  assert_equals: prototype of CSSPseudoElement is not EventTarget expected function "function EventTarget() { [native code] }" but got function "function () { [native code] }"
[FAIL] CSSPseudoElement interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSPseudoElement.prototype is not EventTarget.prototype expected object "[object EventTarget]" but got object "[object Object]"
[FAIL] CSSPseudoElement must be primary interface of beforeElements.item(0)
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] Stringification of beforeElements.item(0)
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] CSSPseudoElement interface: beforeElements.item(0) must inherit property "type" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] CSSPseudoElement interface: beforeElements.item(0) must inherit property "element" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] CSSPseudoElement interface: beforeElements.item(0) must inherit property "parent" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] CSSPseudoElement interface: beforeElements.item(0) must inherit property "pseudo(CSSOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
[FAIL] CSSPseudoElement interface: calling pseudo(CSSOMString) on beforeElements.item(0) with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
Harness: the test ran to completion.
